# src/server/openapi.yaml
openapi: 3.0.3
info:
  title: وسِيطي API
  description: Real Estate CRM & Platform Integration API
  version: 1.0.0
  contact:
    name: Waseety API Support
    email: api@waseety.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.waseety.com/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [villa, apartment, land, commercial]
        category:
          type: string
          enum: [sale, rent]
        city:
          type: string
        price:
          type: number
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        area_m2:
          type: number
...
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
security:
  - bearerAuth: []
paths:
...
  /properties:
    post:
      tags:
        - Properties
      summary: إنشاء عقار جديد
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - type
                - category
                - city
                - price
              properties:
                title:
                  type: string
                  example: "فيلا فاخرة في الياسمين"
                description:
                  type: string
                type:
                  type: string
                  enum: [villa, apartment, land, commercial]
                category:
                  type: string
                  enum: [sale, rent]
                city:
                  type: string
                  example: "الرياض"
                district:
                  type: string
                  example: "الياسمين"
                price:
                  type: number
                  example: 2500000
                bedrooms:
                  type: integer
                bathrooms:
                  type: integer
                area_m2:
                  type: number
                features:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Property'

    get:
      tags:
        - Properties
      summary: قائمة العقارات
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, sold, archived]
        - name: city
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of properties
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Property'
  /properties/{id}/files:
    post:
      tags:
        - Properties
      summary: رفع ملفات العقار
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                file_type:
                  type: string
                  enum: [image, video, document]
      responses:
        '201':
          description: Files uploaded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          uploaded_files:
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: string
                                  format: uuid
                                url:
                                  type: string

  # =======================================================
  # Listings
  # =======================================================
...